+++ crates/router/src/core/payments/operations/payment_response.rs
@@ -1,7 +1,6 @@
 use std::collections::HashMap;
 
 use async_trait::async_trait;
-use common_utils::fp_utils;
 use error_stack::ResultExt;
 use router_derive;
 
@@ -54,9 +53,6 @@ impl<F: Clone> PostUpdateTracker<F, PaymentData<F>, types::PaymentsAuthorizeData
             .mandate_id
             .or_else(|| router_data.request.mandate_id.clone());
 
-        let router_response = router_data.response.clone();
-        let connector = router_data.connector.clone();
-
         payment_data = payment_response_update_tracker(
             db,
             payment_id,
@@ -65,22 +61,6 @@ impl<F: Clone> PostUpdateTracker<F, PaymentData<F>, types::PaymentsAuthorizeData
             storage_scheme,
         )
         .await?;
-
-        router_response.map(|_| ()).or_else(|error_response| {
-            fp_utils::when(
-                !(200..300).contains(&error_response.status_code)
-                    && !(500..=511).contains(&error_response.status_code),
-                || {
-                    Err(errors::ApiErrorResponse::ExternalConnectorError {
-                        code: error_response.code,
-                        message: error_response.message,
-                        connector,
-                        status_code: error_response.status_code,
-                        reason: error_response.reason,
-                    })
-                },
-            )
-        })?;
 
         Ok(payment_data)
     }
@@ -119,9 +99,6 @@ impl<F: Clone> PostUpdateTracker<F, PaymentData<F>, types::PaymentsSessionData>
     where
         F: 'b + Send,
     {
-        let router_response = router_data.response.clone();
-        let connector = router_data.connector.clone();
-
         payment_data = payment_response_update_tracker(
             db,
             payment_id,
@@ -130,16 +107,6 @@ impl<F: Clone> PostUpdateTracker<F, PaymentData<F>, types::PaymentsSessionData>
             storage_scheme,
         )
         .await?;
-
-        router_response.map_err(|error_response| {
-            errors::ApiErrorResponse::ExternalConnectorError {
-                message: error_response.message,
-                code: error_response.code,
-                status_code: error_response.status_code,
-                reason: error_response.reason,
-                connector,
-            }
-        })?;
 
         Ok(payment_data)
     }
@@ -160,9 +127,6 @@ impl<F: Clone> PostUpdateTracker<F, PaymentData<F>, types::PaymentsCaptureData>
     where
         F: 'b + Send,
     {
-        let router_response = router_data.response.clone();
-        let connector = router_data.connector.clone();
-
         payment_data = payment_response_update_tracker(
             db,
             payment_id,
@@ -171,16 +135,6 @@ impl<F: Clone> PostUpdateTracker<F, PaymentData<F>, types::PaymentsCaptureData>
             storage_scheme,
         )
         .await?;
-
-        router_response.map_err(|error_response| {
-            errors::ApiErrorResponse::ExternalConnectorError {
-                message: error_response.message,
-                code: error_response.code,
-                status_code: error_response.status_code,
-                reason: error_response.reason,
-                connector,
-            }
-        })?;
 
         Ok(payment_data)
     }
@@ -200,9 +154,6 @@ impl<F: Clone> PostUpdateTracker<F, PaymentData<F>, types::PaymentsCancelData> f
     where
         F: 'b + Send,
     {
-        let router_response = router_data.response.clone();
-        let connector = router_data.connector.clone();
-
         payment_data = payment_response_update_tracker(
             db,
             payment_id,
@@ -211,16 +162,6 @@ impl<F: Clone> PostUpdateTracker<F, PaymentData<F>, types::PaymentsCancelData> f
             storage_scheme,
         )
         .await?;
-
-        router_response.map_err(|error_response| {
-            errors::ApiErrorResponse::ExternalConnectorError {
-                message: error_response.message,
-                code: error_response.code,
-                status_code: error_response.status_code,
-                reason: error_response.reason,
-                connector,
-            }
-        })?;
 
         Ok(payment_data)
     }
@@ -245,9 +186,6 @@ impl<F: Clone> PostUpdateTracker<F, PaymentData<F>, types::VerifyRequestData> fo
             // .map(api_models::payments::MandateIds::new)
         });
 
-        let router_response = router_data.response.clone();
-        let connector = router_data.connector.clone();
-
         payment_data = payment_response_update_tracker(
             db,
             payment_id,
@@ -256,16 +194,6 @@ impl<F: Clone> PostUpdateTracker<F, PaymentData<F>, types::VerifyRequestData> fo
             storage_scheme,
         )
         .await?;
-
-        router_response.map_err(|error_response| {
-            errors::ApiErrorResponse::ExternalConnectorError {
-                message: error_response.message,
-                code: error_response.code,
-                status_code: error_response.status_code,
-                reason: error_response.reason,
-                connector,
-            }
-        })?;
 
         Ok(payment_data)
     }