+++ crates/router/src/services/api.rs
@@ -298,24 +298,30 @@ where
                     match response {
                         Ok(body) => {
                             let response = match body {
-                                Ok(body) => connector_integration
-                                    .handle_response(req, body)
-                                    .map_err(|error| {
-                                        if error.current_context()
-                                        == &errors::ConnectorError::ResponseDeserializationFailed
-                                    {
-                                        metrics::RESPONSE_DESERIALIZATION_FAILURE.add(
-                                            &metrics::CONTEXT,
-                                            1,
-                                            &[metrics::request::add_attributes(
-                                                "connector",
-                                                req.connector.to_string(),
-                                            )],
-                                        )
-                                    }
-                                        error
-                                    })?,
+                                Ok(body) => {
+                                    let connector_http_status_code = Some(body.status_code);
+                                    let mut data = connector_integration
+                                        .handle_response(req, body)
+                                        .map_err(|error| {
+                                            if error.current_context()
+                                            == &errors::ConnectorError::ResponseDeserializationFailed
+                                        {
+                                            metrics::RESPONSE_DESERIALIZATION_FAILURE.add(
+                                                &metrics::CONTEXT,
+                                                1,
+                                                &[metrics::request::add_attributes(
+                                                    "connector",
+                                                    req.connector.to_string(),
+                                                )],
+                                            )
+                                        }
+                                            error
+                                        })?;
+                                    data.connector_http_status_code = connector_http_status_code;
+                                    data
+                                }
                                 Err(body) => {
+                                    router_data.connector_http_status_code = Some(body.status_code);
                                     metrics::CONNECTOR_ERROR_RESPONSE_COUNT.add(
                                         &metrics::CONTEXT,
                                         1,
@@ -541,6 +547,7 @@ pub enum ApplicationResponse<R> {
     JsonForRedirection(api::RedirectionResponse),
     Form(Box<RedirectionFormData>),
     FileData((Vec<u8>, mime::Mime)),
+    JsonWithHeaders((R, Vec<(String, String)>)),
 }
 
 #[derive(Debug, Eq, PartialEq)]
@@ -717,6 +724,18 @@ where
         )
         .respond_to(request)
         .map_into_boxed_body(),
+        Ok(ApplicationResponse::JsonWithHeaders((response, headers))) => {
+            match serde_json::to_string(&response) {
+                Ok(res) => http_response_json_with_headers(res, headers),
+                Err(_) => http_response_err(
+                    r#"{
+                        "error": {
+                            "message": "Error serializing response from connector"
+                        }
+                    }"#,
+                ),
+            }
+        }
         Err(error) => log_and_return_error_response(error),
     };
 
@@ -778,6 +797,19 @@ impl EmbedError for Report<api_models::errors::types::ApiErrorResponse> {
 
 pub fn http_response_json<T: body::MessageBody + 'static>(response: T) -> HttpResponse {
     HttpResponse::Ok()
+        .content_type(mime::APPLICATION_JSON)
+        .body(response)
+}
+
+pub fn http_response_json_with_headers<T: body::MessageBody + 'static>(
+    response: T,
+    headers: Vec<(String, String)>,
+) -> HttpResponse {
+    let mut response_builder = HttpResponse::Ok();
+    for (name, value) in headers {
+        response_builder.append_header((name, value));
+    }
+    response_builder
         .content_type(mime::APPLICATION_JSON)
         .body(response)
 }